refactor: use plugin to support hooks
